+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - \_process_item:  Private function for process to decrease complexity.
 
 ### Changed
+- process:  Replaced section of code with call to \_process_other function.
 - process:  Replaced section of code with call to \_process_item function.
 - transfer_file:  Refactored if statements.
 - transfer_file:  Removed check for job log.